Clean up plonky3 backend test helpers
